ai-service/requirements.txt
flask
flask-cors
flask-limiter
requests
python-dotenv
redis